@@ -36,17 +36,12 @@ public class SimpleGame {
 
         //Schedule a job for the event-dispatching thread:
         //creating and showing this application's GUI.
-        javax.swing.SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                jframe = new SimpleGameJFrame(
-                        SimpleGameConfig.getInstance().getGameTitle(),
-                        SimpleGameConfig.getInstance().getGameWidth(),
-                        SimpleGameConfig.getInstance().getGameHeight(),
-                        SimpleGameConfig.getInstance().getGameTimerDelay(),
-                        SimpleGameConfig.getInstance().getZoom());
-            }
-        });
+        javax.swing.SwingUtilities.invokeLater(() -> jframe = new SimpleGameJFrame(
+                SimpleGameConfig.getInstance().getGameTitle(),
+                SimpleGameConfig.getInstance().getGameWidth(),
+                SimpleGameConfig.getInstance().getGameHeight(),
+                SimpleGameConfig.getInstance().getGameTimerDelay(),
+                SimpleGameConfig.getInstance().getZoom()));
     }
 
     /**
