@@ -19,12 +19,9 @@ public class StartMenuJill1Handler extends AbstractChangeLevel {
      * Default constructor of level.
      *
      * @throws IOException if missing file
-     * @throws ClassNotFoundException if missing class must be load
-     * @throws IllegalAccessException if trouble when class must be load
-     * @throws InstantiationException if trouble when class must be load
+     * @throws ReflectiveOperationException if missing class must be load
      */
-    public StartMenuJill1Handler() throws IOException, ClassNotFoundException,
-        IllegalAccessException, InstantiationException {
+    public StartMenuJill1Handler() throws IOException, ReflectiveOperationException {
         super(new JillLevelConfiguration("JILL1.SHA", "INTRO.JN1", "JILL1.VCL",
             "JILL1.CFG", "JN1", StartMenuJill1Handler.class, false));
         centerScreen();
